+++ ampd/src/grpc/blockchain_service.rs
@@ -1,16 +1,13 @@
 use std::fmt::Debug;
 use std::pin::Pin;
-use std::sync::Arc;
 
 use ampd_proto::blockchain_service_server::BlockchainService;
 use ampd_proto::{
-    AddressRequest, AddressResponse, BroadcastRequest, BroadcastResponse, ContractStateRequest,
-    ContractStateResponse, ContractsRequest, ContractsResponse, SubscribeRequest,
-    SubscribeResponse,
+    ContractStateRequest, ContractStateResponse, ContractsRequest, ContractsResponse,
+    SubscribeRequest, SubscribeResponse,
 };
 use async_trait::async_trait;
-use axelar_wasm_std::FnExt;
-use futures::{Stream, TryFutureExt, TryStreamExt};
+use futures::{Stream, TryStreamExt};
 use tokio_stream::StreamExt;
 use tonic::{Request, Response, Status};
 use tracing::instrument;
@@ -18,7 +15,7 @@ use typed_builder::TypedBuilder;
 
 use crate::grpc::reqs::Validate;
 use crate::grpc::status;
-use crate::{broadcast, cosmos, event_sub};
+use crate::{cosmos, event_sub};
 
 #[derive(Debug, TypedBuilder)]
 pub struct Service<E, C>
@@ -27,7 +24,6 @@ where
     C: cosmos::CosmosClient,
 {
     event_sub: E,
-    msg_queue_client: broadcast::MsgQueueClient<C>,
     cosmos_client: C,
 }
 
@@ -64,28 +60,6 @@ where
     }
 
     #[instrument]
-    async fn broadcast(
-        &self,
-        req: Request<BroadcastRequest>,
-    ) -> Result<Response<BroadcastResponse>, Status> {
-        let msg = req
-            .validate()
-            .inspect_err(status::log("invalid broadcast request"))
-            .map_err(status::StatusExt::into_status)?;
-
-        self.msg_queue_client
-            .clone()
-            .enqueue(msg)
-            .map_err(Arc::new)
-            .and_then(|rx| rx)
-            .await
-            .map(|(tx_hash, index)| BroadcastResponse { tx_hash, index })
-            .map(Response::new)
-            .inspect_err(|err| err.as_ref().then(status::log("message broadcast error")))
-            .map_err(|err| status::StatusExt::into_status(err.as_ref()))
-    }
-
-    #[instrument]
     async fn contract_state(
         &self,
         req: Request<ContractStateRequest>,
@@ -103,15 +77,6 @@ where
             .map_err(status::StatusExt::into_status)
     }
 
-    async fn address(
-        &self,
-        _req: Request<AddressRequest>,
-    ) -> Result<Response<AddressResponse>, Status> {
-        Ok(Response::new(AddressResponse {
-            address: self.msg_queue_client.address().to_string(),
-        }))
-    }
-
     async fn contracts(
         &self,
         _req: Request<ContractsRequest>,
@@ -124,109 +89,22 @@ where
 
 #[cfg(test)]
 mod tests {
-    use std::sync::Arc;
-    use std::time::Duration;
-
-    use axelar_wasm_std::nonempty;
-    use cosmos_sdk_proto::cosmos::bank::v1beta1::{QueryBalanceRequest, QueryBalanceResponse};
-    use cosmos_sdk_proto::cosmos::base::v1beta1::Coin;
-    use cosmrs::proto::cosmos::auth::v1beta1::{BaseAccount, QueryAccountResponse};
-    use cosmrs::proto::cosmos::base::abci::v1beta1::GasInfo;
-    use cosmrs::proto::cosmos::tx::v1beta1::SimulateResponse;
     use cosmrs::proto::cosmwasm::wasm::v1::{
         QuerySmartContractStateRequest, QuerySmartContractStateResponse,
     };
-    use cosmrs::{Any, Gas};
     use error_stack::report;
     use events::{self, Event};
-    use futures::future::join_all;
     use futures::{stream, StreamExt};
-    use mockall::{predicate, Sequence};
+    use mockall::predicate;
     use report::ErrorExt;
     use tokio_stream::wrappers::errors::BroadcastStreamRecvError;
     use tonic::{Code, Request};
 
     use super::*;
-    use crate::broadcast::DecCoin;
     use crate::cosmos::MockCosmosClient;
     use crate::event_sub::{self, MockEventSub};
-    use crate::types::{random_cosmos_public_key, TMAddress};
+    use crate::types::TMAddress;
     use crate::PREFIX;
-
-    const GAS_CAP: Gas = 10000;
-
-    async fn setup(
-        mock_event_sub: MockEventSub,
-        mut broadcaster_mock_cosmos_client: MockCosmosClient,
-        mock_cosmos_client: MockCosmosClient,
-    ) -> (
-        Service<MockEventSub, MockCosmosClient>,
-        impl Stream<Item = nonempty::Vec<broadcast::QueueMsg>>,
-    ) {
-        let pub_key = random_cosmos_public_key();
-        let address: TMAddress = pub_key.account_id(PREFIX).unwrap().into();
-        let gas_adjustment = 1.5;
-        let gas_price_amount = 0.025;
-        let gas_price_denom = "uaxl";
-
-        let base_account = BaseAccount {
-            address: address.to_string(),
-            pub_key: None,
-            account_number: 42,
-            sequence: 10,
-        };
-
-        let mut seq = Sequence::new();
-        broadcaster_mock_cosmos_client
-            .expect_account()
-            .once()
-            .in_sequence(&mut seq)
-            .return_once(move |_| {
-                Ok(QueryAccountResponse {
-                    account: Some(Any::from_msg(&base_account).unwrap()),
-                })
-            });
-
-        broadcaster_mock_cosmos_client
-            .expect_balance()
-            .once()
-            .with(predicate::eq(QueryBalanceRequest {
-                address: address.to_string(),
-                denom: gas_price_denom.to_string(),
-            }))
-            .in_sequence(&mut seq)
-            .return_once(move |_| {
-                Ok(QueryBalanceResponse {
-                    balance: Some(Coin {
-                        denom: gas_price_denom.to_string(),
-                        amount: "1000000".to_string(),
-                    }),
-                })
-            });
-
-        let broadcaster = broadcast::Broadcaster::builder()
-            .client(broadcaster_mock_cosmos_client)
-            .chain_id("chain_id".try_into().unwrap())
-            .pub_key(pub_key)
-            .gas_adjustment(gas_adjustment)
-            .gas_price(DecCoin::new(gas_price_amount, gas_price_denom).unwrap())
-            .build()
-            .await
-            .unwrap();
-        let (msg_queue, msg_queue_client) = broadcast::MsgQueue::new_msg_queue_and_client(
-            broadcaster,
-            100,
-            GAS_CAP,
-            Duration::from_secs(1),
-        );
-        let service = Service::builder()
-            .event_sub(mock_event_sub)
-            .msg_queue_client(msg_queue_client)
-            .cosmos_client(mock_cosmos_client)
-            .build();
-
-        (service, msg_queue)
-    }
 
     #[tokio::test]
     async fn subscribe_should_stream_events_successfully() {
@@ -242,12 +120,11 @@ mod tests {
             .expect_subscribe()
             .return_once(move || stream::iter(events.into_iter().map(Result::Ok)).boxed());
 
-        let (service, _) = setup(
-            mock_event_sub,
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(MockCosmosClient::new())
+            .build();
+
         let res = service
             .subscribe(subscribe_req(vec![], true))
             .await
@@ -263,12 +140,10 @@ mod tests {
 
     #[tokio::test]
     async fn subscribe_should_return_error_if_any_filter_is_invalid() {
-        let (service, _) = setup(
-            MockEventSub::new(),
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(MockEventSub::new())
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let res = service
             .subscribe(subscribe_req(
                 vec![ampd_proto::EventFilter::default()],
@@ -296,12 +171,10 @@ mod tests {
             tokio_stream::once(Err(report!(event_sub::Error::LatestBlockQuery))).boxed()
         });
 
-        let (service, _) = setup(
-            mock_event_sub,
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let res = service
             .subscribe(subscribe_req(
                 vec![ampd_proto::EventFilter {
@@ -329,12 +202,10 @@ mod tests {
             .boxed()
         });
 
-        let (service, _) = setup(
-            mock_event_sub,
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let res = service
             .subscribe(subscribe_req(
                 vec![ampd_proto::EventFilter {
@@ -362,12 +233,10 @@ mod tests {
             .boxed()
         });
 
-        let (service, _) = setup(
-            mock_event_sub,
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let res = service
             .subscribe(subscribe_req(
                 vec![ampd_proto::EventFilter {
@@ -392,12 +261,10 @@ mod tests {
             tokio_stream::once(Err(BroadcastStreamRecvError::Lagged(10).into_report())).boxed()
         });
 
-        let (service, _) = setup(
-            mock_event_sub,
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let res = service
             .subscribe(subscribe_req(
                 vec![ampd_proto::EventFilter {
@@ -433,12 +300,10 @@ mod tests {
             r#type: "event_type_2".to_string(),
             ..Default::default()
         };
-        let (service, _) = setup(
-            mock_event_sub,
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let res = service
             .subscribe(subscribe_req(vec![filter], false))
             .await
@@ -476,12 +341,10 @@ mod tests {
             r#type: "test_event".to_string(),
             contract: expected.contract_address().unwrap().to_string(),
         };
-        let (service, _) = setup(
-            mock_event_sub,
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let res = service
             .subscribe(subscribe_req(vec![filter], false))
             .await
@@ -507,12 +370,10 @@ mod tests {
             .expect_subscribe()
             .return_once(move || stream::iter(events.into_iter().map(Result::Ok)).boxed());
 
-        let (service, _) = setup(
-            mock_event_sub,
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let res = service
             .subscribe(subscribe_req(vec![], false))
             .await
@@ -562,12 +423,10 @@ mod tests {
             contract: expected[1].contract_address().unwrap().to_string(),
             ..Default::default()
         };
-        let (service, _) = setup(
-            mock_event_sub,
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let res = service
             .subscribe(subscribe_req(vec![filter_1, filter_2], false))
             .await
@@ -582,147 +441,11 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn broadcast_should_return_error_if_req_is_invalid() {
-        let (service, _) = setup(
-            MockEventSub::new(),
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
-        let res = service.broadcast(broadcast_req(None)).await;
-        assert!(res.is_err_and(|status| status.code() == Code::InvalidArgument));
-    }
-
-    #[tokio::test]
-    async fn broadcast_should_return_error_if_enqueue_failed() {
-        let mut mock_cosmos_client = MockCosmosClient::new();
-        mock_cosmos_client.expect_clone().return_once(|| {
-            let mut mock_cosmos_client = MockCosmosClient::new();
-            mock_cosmos_client
-                .expect_simulate()
-                .return_once(|_| Err(Status::internal("simulate error").into_report()));
-
-            mock_cosmos_client
-        });
-
-        let (service, _) = setup(
-            MockEventSub::new(),
-            mock_cosmos_client,
-            MockCosmosClient::new(),
-        )
-        .await;
-        let res = service.broadcast(broadcast_req(Some(dummy_msg()))).await;
-        assert!(res.is_err_and(|status| status.code() == Code::InvalidArgument));
-    }
-
-    #[tokio::test]
-    async fn broadcast_should_return_error_if_broadcast_failed() {
-        let mut mock_cosmos_client = MockCosmosClient::new();
-        mock_cosmos_client.expect_clone().return_once(|| {
-            let mut mock_cosmos_client = MockCosmosClient::new();
-            mock_cosmos_client.expect_simulate().return_once(|_| {
-                Ok(SimulateResponse {
-                    gas_info: Some(GasInfo {
-                        gas_wanted: GAS_CAP + 1,
-                        gas_used: GAS_CAP + 1,
-                    }),
-                    result: None,
-                })
-            });
-
-            mock_cosmos_client
-        });
-
-        let (service, mut msg_queue) = setup(
-            MockEventSub::new(),
-            mock_cosmos_client,
-            MockCosmosClient::new(),
-        )
-        .await;
-        tokio::spawn(async move { while msg_queue.next().await.is_some() {} });
-        let res = service.broadcast(broadcast_req(Some(dummy_msg()))).await;
-        assert!(res.is_err_and(|status| status.code() == Code::InvalidArgument));
-    }
-
-    #[tokio::test]
-    async fn broadcast_should_return_tx_hash_and_index() {
-        let tx_hash = "0x7cedbb3799cd99636045c84c5c55aef8a138f107ac8ba53a08cad1070ba4385b";
-        let msg_count = 10;
-        let mut mock_cosmos_client = MockCosmosClient::new();
-        mock_cosmos_client
-            .expect_clone()
-            .times(msg_count)
-            .returning(move || {
-                let mut mock_cosmos_client = MockCosmosClient::new();
-                mock_cosmos_client.expect_simulate().return_once(move |_| {
-                    Ok(SimulateResponse {
-                        gas_info: Some(GasInfo {
-                            gas_wanted: GAS_CAP / msg_count as u64,
-                            gas_used: GAS_CAP / msg_count as u64,
-                        }),
-                        result: None,
-                    })
-                });
-
-                mock_cosmos_client
-            });
-
-        let (service, mut msg_queue) = setup(
-            MockEventSub::new(),
-            mock_cosmos_client,
-            MockCosmosClient::new(),
-        )
-        .await;
-        let service = Arc::new(service);
-        let handles = join_all(
-            (0..msg_count)
-                .map(|_| {
-                    let service = service.clone();
-
-                    tokio::spawn(async move {
-                        let res = service
-                            .broadcast(broadcast_req(Some(dummy_msg())))
-                            .await
-                            .unwrap()
-                            .into_inner();
-
-                        (res.tx_hash, res.index)
-                    })
-                })
-                .collect::<Vec<_>>(),
-        );
-
-        let msgs: Vec<_> = msg_queue.next().await.unwrap().into();
-        assert_eq!(msgs.len(), msg_count);
-        for (i, msg) in msgs.into_iter().enumerate() {
-            assert_eq!(msg.gas, GAS_CAP / msg_count as u64);
-            msg.tx_res_callback
-                .send(Ok((tx_hash.to_string(), i as u64)))
-                .unwrap();
-        }
-
-        let mut results = handles.await;
-        results.sort_by(|result_a, result_b| {
-            let result_a = result_a.as_ref().unwrap();
-            let result_b = result_b.as_ref().unwrap();
-
-            result_a.1.cmp(&result_b.1)
-        });
-        for (i, result) in results.into_iter().enumerate() {
-            let (tx_hash, index) = result.unwrap();
-            assert_eq!(tx_hash, tx_hash.to_string());
-            assert_eq!(index, i as u64);
-        }
-    }
-
-    #[tokio::test]
     async fn contract_state_should_return_error_if_req_is_invalid() {
-        let (service, _) = setup(
-            MockEventSub::new(),
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(MockEventSub::new())
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let req = Request::new(ContractStateRequest {
             contract: "invalid_address".to_string(),
             query: serde_json::to_vec(&serde_json::json!({"get_config": {}})).unwrap(),
@@ -735,12 +458,10 @@ mod tests {
     #[tokio::test]
     async fn contract_state_should_return_error_if_empty_query() {
         let address = TMAddress::random(PREFIX);
-        let (service, _) = setup(
-            MockEventSub::new(),
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(MockEventSub::new())
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let req = Request::new(ContractStateRequest {
             contract: address.to_string(),
             query: vec![],
@@ -753,12 +474,10 @@ mod tests {
     #[tokio::test]
     async fn contract_state_should_return_error_if_invalid_json() {
         let address = TMAddress::random(PREFIX);
-        let (service, _) = setup(
-            MockEventSub::new(),
-            MockCosmosClient::new(),
-            MockCosmosClient::new(),
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(MockEventSub::new())
+            .cosmos_client(MockCosmosClient::new())
+            .build();
         let req = Request::new(ContractStateRequest {
             contract: address.to_string(),
             query: vec![1, 2, 3], // invalid json
@@ -794,12 +513,10 @@ mod tests {
             mock
         });
 
-        let (service, _) = setup(
-            MockEventSub::new(),
-            MockCosmosClient::new(),
-            mock_cosmos_client,
-        )
-        .await;
+        let service = Service::builder()
+            .event_sub(MockEventSub::new())
+            .cosmos_client(mock_cosmos_client)
+            .build();
         let req = Request::new(ContractStateRequest {
             contract: address_str,
             query: query_bytes,
@@ -839,12 +556,11 @@ mod tests {
             mock
         });
 
-        let (service, _) = setup(
-            MockEventSub::new(),
-            MockCosmosClient::new(),
-            mock_cosmos_client,
-        )
-        .await;
+        let mock_event_sub = MockEventSub::new();
+        let service = Service::builder()
+            .event_sub(mock_event_sub)
+            .cosmos_client(mock_cosmos_client)
+            .build();
         let req = Request::new(ContractStateRequest {
             contract: address_str,
             query: query_bytes,
@@ -852,63 +568,6 @@ mod tests {
         let res = service.contract_state(req).await.unwrap().into_inner();
 
         assert_eq!(res.result, result);
-    }
-
-    #[tokio::test]
-    async fn address_should_return_msg_queue_client_address() {
-        let pub_key = random_cosmos_public_key();
-        let expected_address: TMAddress = pub_key.account_id(PREFIX).unwrap().into();
-        let base_account = BaseAccount {
-            address: expected_address.to_string(),
-            pub_key: None,
-            account_number: 42,
-            sequence: 10,
-        };
-        let gas_adjustment = 1.5;
-        let gas_price_amount = 0.025;
-        let gas_price_denom = "uaxl";
-
-        let mut mock_cosmos_client = MockCosmosClient::new();
-        mock_cosmos_client.expect_account().return_once(move |_| {
-            Ok(QueryAccountResponse {
-                account: Some(Any::from_msg(&base_account).unwrap()),
-            })
-        });
-        mock_cosmos_client.expect_balance().return_once(move |_| {
-            Ok(QueryBalanceResponse {
-                balance: Some(Coin {
-                    denom: "uaxl".to_string(),
-                    amount: "1000000".to_string(),
-                }),
-            })
-        });
-        let broadcaster = broadcast::Broadcaster::builder()
-            .client(mock_cosmos_client)
-            .chain_id("chain-id".parse().unwrap())
-            .pub_key(pub_key)
-            .gas_adjustment(gas_adjustment)
-            .gas_price(DecCoin::new(gas_price_amount, gas_price_denom).unwrap())
-            .build()
-            .await
-            .unwrap();
-
-        let (_, msg_queue_client) = broadcast::MsgQueue::new_msg_queue_and_client(
-            broadcaster,
-            10,
-            1000u64,
-            Duration::from_secs(1),
-        );
-
-        let service = Service::builder()
-            .event_sub(MockEventSub::new())
-            .msg_queue_client(msg_queue_client)
-            .cosmos_client(MockCosmosClient::new())
-            .build();
-
-        let req = Request::new(AddressRequest {});
-        let res = service.address(req).await.unwrap().into_inner();
-
-        assert_eq!(res.address, expected_address.to_string());
     }
 
     fn subscribe_req(
@@ -919,10 +578,6 @@ mod tests {
             filters,
             include_block_begin_end,
         })
-    }
-
-    fn broadcast_req(msg: Option<Any>) -> Request<BroadcastRequest> {
-        Request::new(BroadcastRequest { msg })
     }
 
     fn block_begin_event(height: u64) -> Event {
@@ -955,13 +610,6 @@ mod tests {
                     )
                 })
                 .collect(),
-        }
-    }
-
-    fn dummy_msg() -> Any {
-        Any {
-            type_url: "/cosmos.bank.v1beta1.MsgSend".to_string(),
-            value: vec![1, 2, 3, 4],
         }
     }
 }
